Parameters to set timeout by request method, higher default for POST and DELETE
It's not a good idea to set low default timeout for POST and DELETE requests
as these may take a long time, depending on compute resource and how
deployment works in given environment.

10 minute timeout for POSTs and DELETEs should be safe enough. User can set
these timeouts to None (default timeout) or 0 (no timeout for given method).

It's a pitty that these Foreman resources are not asynchronous.