Raise ObjectNotFound exception on 404 except for GET requests
We used to always return an empty list if we got a HTTP error 404 in response
to our requests, ignoring any error message possibly returned by Foreman.

For GET requests this is desirable because trying to GET a non-existant
resource (eg. using a ID specified by the user) is perfectly normal and does
not justify raising an exception. For other requests (especially POST) it is
and thus we now raise the ObjectNotFound exception (which so far had not been
used anywhere) for anything but GET requests when a 404 is encountered.